+++ src/mon/vision/enhance/llie/llflow/utils/util.py
@@ -212,8 +212,8 @@ def save_img(img, img_path, mode='RGB'):
 
 def calculate_psnr(img1, img2):
     # img1 and img2 have range [0, 255]
-    img1 = img1.astype(np.float64)
-    img2 = img2.astype(np.float64)
+    img1 = img1.astype(np.float32)
+    img2 = img2.astype(np.float32)
     mse = np.mean((img1 - img2) ** 2)
     if mse == 0:
         return float('inf')
